pyproject.toml
[project]
name = "utilshub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "peek-python>=25.0.11",
    "pyside6>=6.8.2.1",
]
